+++ integration_test/cases/repo.exs
@@ -269,15 +269,19 @@ defmodule Ecto.Integration.RepoTest do
 
   test "optimistic locking in update/delete operations" do
     import Ecto.Changeset, only: [cast: 3, optimistic_lock: 2]
-    base_post = TestRepo.insert!(%Comment{})
+    base_comment = TestRepo.insert!(%Comment{})
 
     changeset_ok =
-      base_post
+      base_comment
       |> cast(%{"text" => "foo.bar"}, ~w(text)a)
       |> optimistic_lock(:lock_version)
     TestRepo.update!(changeset_ok)
 
-    changeset_stale = optimistic_lock(base_post, :lock_version)
+    changeset_stale =
+      base_comment
+      |> cast(%{"text" => "foo.bat"}, ~w(text)a)
+      |> optimistic_lock(:lock_version)
+
     assert_raise Ecto.StaleEntryError, fn -> TestRepo.update!(changeset_stale) end
     assert_raise Ecto.StaleEntryError, fn -> TestRepo.delete!(changeset_stale) end
   end
@@ -285,7 +289,7 @@ defmodule Ecto.Integration.RepoTest do
   test "optimistic locking in update operation with nil field" do
     import Ecto.Changeset, only: [cast: 3, optimistic_lock: 3]
 
-    base_post =
+    base_comment =
       %Comment{}
       |> cast(%{lock_version: nil}, [:lock_version])
       |> TestRepo.insert!()
@@ -297,7 +301,7 @@ defmodule Ecto.Integration.RepoTest do
       end
 
     changeset_ok =
-      base_post
+      base_comment
       |> cast(%{"text" => "foo.bar"}, ~w(text)a)
       |> optimistic_lock(:lock_version, incrementer)
 
@@ -309,7 +313,7 @@ defmodule Ecto.Integration.RepoTest do
   test "optimistic locking in delete operation with nil field" do
     import Ecto.Changeset, only: [cast: 3, optimistic_lock: 3]
 
-    base_post =
+    base_comment =
       %Comment{}
       |> cast(%{lock_version: nil}, [:lock_version])
       |> TestRepo.insert!()
@@ -320,10 +324,10 @@ defmodule Ecto.Integration.RepoTest do
         old_value -> old_value + 1
       end
 
-    changeset_ok = optimistic_lock(base_post, :lock_version, incrementer)
+    changeset_ok = optimistic_lock(base_comment, :lock_version, incrementer)
     TestRepo.delete!(changeset_ok)
 
-    refute TestRepo.get(Comment, base_post.id)
+    refute TestRepo.get(Comment, base_comment.id)
   end
 
   @tag :unique_constraint