+++ test/ecto/changeset_test.exs
@@ -1474,32 +1474,27 @@ defmodule Ecto.ChangesetTest do
     changeset = changeset(%{}) |> optimistic_lock(:upvotes)
     assert changeset.filters == %{upvotes: 0}
     assert changeset.changes == %{}
-    assert changeset.repo_opts == [force: true]
     assert prepared_changes(changeset) == %{upvotes: 1}
 
     changeset = changeset(%Post{upvotes: 2}, %{upvotes: 1}) |> optimistic_lock(:upvotes)
     assert changeset.filters == %{upvotes: 1}
     assert changeset.changes == %{upvotes: 1}
-    assert changeset.repo_opts == [force: true]
     assert prepared_changes(changeset) == %{upvotes: 2}
 
     # Assert default increment will rollover to 1 when the current one is equal or graeter than 2_147_483_647
     changeset = changeset(%Post{upvotes: 2_147_483_647}, %{}) |> optimistic_lock(:upvotes)
     assert changeset.filters == %{upvotes: 2_147_483_647}
     assert changeset.changes == %{}
-    assert changeset.repo_opts == [force: true]
     assert prepared_changes(changeset) == %{upvotes: 1}
 
     changeset = changeset(%Post{upvotes: 3_147_483_647}, %{}) |> optimistic_lock(:upvotes)
     assert changeset.filters == %{upvotes: 3_147_483_647}
     assert changeset.changes == %{}
-    assert changeset.repo_opts == [force: true]
     assert prepared_changes(changeset) == %{upvotes: 1}
 
     changeset = changeset(%Post{upvotes: 2_147_483_647}, %{upvotes: 2_147_483_648}) |> optimistic_lock(:upvotes)
     assert changeset.filters == %{upvotes: 2_147_483_648}
     assert changeset.changes == %{upvotes: 2_147_483_648}
-    assert changeset.repo_opts == [force: true]
     assert prepared_changes(changeset) == %{upvotes: 1}
  end
 
@@ -1507,7 +1502,6 @@ defmodule Ecto.ChangesetTest do
     changeset = %Post{} |> optimistic_lock(:upvotes)
     assert changeset.filters == %{upvotes: 0}
     assert changeset.changes == %{}
-    assert changeset.repo_opts == [force: true]
     assert prepared_changes(changeset) == %{upvotes: 1}
   end
 
@@ -1515,7 +1509,6 @@ defmodule Ecto.ChangesetTest do
     changeset = %Post{} |> optimistic_lock(:upvotes, &(&1 - 1))
     assert changeset.filters == %{upvotes: 0}
     assert changeset.changes == %{}
-    assert changeset.repo_opts == [force: true]
     assert prepared_changes(changeset) == %{upvotes: -1}
   end